+++ store/index.js
@@ -1,19 +1,31 @@
 /* eslint-disable no-console */
 
 const CAT_SERVICE = 'https://cataas.com/cat'
+// https://icanhazdadjoke.com/api
+const DAD_QUOTE = 'https://icanhazdadjoke.com/'
 
 const FileSaver = require('file-saver')
 
 const getDefaultState = () => {
   return {
-    quote: {}
+    quote: {},
+    pix: {},
+    quoteLocation: 'above',
+    imageType: 'gif',
+    theme: '',
+    pixId: '60ef3f0151a2ca0011c7455d'
   }
 }
 
 export const state = () => getDefaultState()
 
 export const getters = {
-  getQuote: state => state.documentationTypes
+  getQuote: state => state.quote,
+  getPix: state => state.pix,
+  getQuoteLocation: state => state.quoteLocation,
+  getPixId: state => state.pixId,
+  getImageType: state => state.imageType,
+  getTheme: state => state.theme
   // getTypeByCode: state => (code) => {
   //   return state.documentationTypes.find(value => value.code === code).description
   // }
@@ -22,6 +34,18 @@ export const getters = {
 export const mutations = {
   setQuote (state, value) {
     state.quote = value
+  },
+  setPix (state, value) {
+    state.pix = value
+  },
+  setQuoteLocation (state, value) {
+    state.quoteLocation = value
+  },
+  setPixId (state, value) {
+    state.pixId = value
+  },
+  setTheme (state, value) {
+    state.theme = value
   }
 }
 
@@ -36,23 +60,20 @@ export const mutations = {
  * @param {*} commitTo the state object to update
  * @returns a success or error message
  */
-async function getAxios ({ commit }, ax, url, params, title, message, commitTo) {
+async function getAxios ({ commit }, ax, url, params, config, title, message, commitTo) {
   let msg, data
   try {
-    const config = {
-      params
-    }
-    data = await ax.$get(url, config)
+    data = await ax.$get(url, params, config)
     console.debug('config: ', config)
     console.info(title + '  DATA from get formdata')
     console.debug(data)
 
-    if (data.length > 0) {
+    if (data.length > 0 || data.status === 200) {
       commit(commitTo, data)
-      console.debug('# of items returned from axios call: ' + data.length)
+      console.info('# of items returned from axios call: ' + data.length)
       msg = 'success'
     } else {
-      msg = 'Problem getting ' + title
+      msg = 'Problem getting ' + title + ' from ' + data
     }
   } catch (err) {
     console.error('error ' + title)
@@ -74,11 +95,11 @@ async function getAxios ({ commit }, ax, url, params, title, message, commitTo) 
  * @param {*} commitTo the state object to update
  * @returns a success or error message
  */
-async function processAxios ({ commit }, ax, url, title, message, commitTo) {
-  console.debug(title + ' 2 url: ' + url)
+// async function processAxios ({ commit }, ax, url, title, message, commitTo) {
+//   console.debug(title + ' 2 url: ' + url)
 
-  return await getAxios({ commit }, ax, url, null, title, message, commitTo)
-}
+//   return await getAxios({ commit }, ax, url, null, null, title, message, commitTo)
+// }
 
 export const actions = {
   getMenuItems () {
@@ -94,8 +115,38 @@ export const actions = {
 
   async getCatPix ({ commit, state }, params) {
     const url = CAT_SERVICE
+    const result = await this.$axios.get(url, {})
+    let newCat = {}
+    const imageUrl = URL.createObjectURL(result)
+    if (imageUrl) {
+      // Update your store with an object like this:
+      newCat = { id: '123', url: imageUrl }
+    }
+    commit('setPix', newCat)
+    // return await processAxios({ commit }, this.$axios, url, params, 'Cat As Service', 'Cat Service', 'setPix')
+  },
 
-    return await processAxios({ commit }, this.$axios, url, params, 'Cat As Service', 'Cat Service', 'setDocumentList')
+  async getDadQuote ({ commit, state }, params) {
+    const url = DAD_QUOTE
+    // 'User-Agent': 'Catty (https://github.com/jimmerydad/cat-quotes)'
+    const config = {
+      headers: {
+        Accept: 'application/json',
+        'Content-Type': 'application/json'
+      }
+    }
+    const result = await getAxios({ commit }, this.$axios, url, config, null, 'Dad Quote', 'Dad Quote Service', 'setQuote')
+    console.log('result: ', result)
+    // Accept: 'application/json',
+    // 'User-Agent': 'Catty (https://github.com/jimmerydad/cat-quotes)',
+    const result2 = await this.$axios.get(url, {
+      headers: {
+        Accept: 'application/json',
+        'Content-Type': 'application/json'
+      }
+    })
+    console.log('result2: ', result2)
+    return result
   },
 
   // async updateUserProfile ({ commit, state }, up) {